EquityEye/EquityEye-backend/requirements.txt
django
mysqlclient
djangorestframework
django-cors-headers
djangorestframework-simplejwt
requests
openai